backend-fast/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
pillow==10.1.0
openai==1.3.7
google-generativeai==0.3.2
boto3==1.34.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pydantic[email]
flower>=2.0.0
requests